BackEnd Изменение структуры контроллеров -Разбиение контроллеров на логически связанные части -Изменение запросов:
Администратор:

Уведомления
admin/add_notification -> admin/notifications/add
admin/delete_notification -> admin/notifications/delete
admin/get_active_notifications -> admin/notifications/get_active

Пользователь без привелегий:

Уведомления
dashboard/get_notifications_by_group -> dashboard/notifications/get_for_group

Группы
dashboard/get_list_groups -> dashboard/groups/get_list
dashboard/get_list_group_without_grade -> dashboard/groups/get_list_without_grade
dashboard/get_filtered_groups -> dashboard/groups/filter_apply

Преподаватели
dashboard/get_professor_state -> dashboard/professors/get_professor_state
dashboard/get_professor_timetable -> dashboard/professors/get_professor_timetable
dashboard/get_list_professors -> dashboard/professors/get_list

Расписание
dashboard/get_actual_dashboard -> dashboard/timetable/get_actual
dashboard/get_week_dashboard -> dashboard/timetable/get_week
dashboard/get_lesson_info -> dashboard/timetable/get_lesson_info
dashboard/get_working_days_group_for_month -> dashboard/timetable/get_working_days_for_month

Подразделения
dashboard/get_faculty_list -> dashboard/units/get_faculty_list
dashboard/get_specializations_list -> dashboard/units/get_spec_list
dashboard/get_deparmetns_list -> dashboard/units/get_depart_list

Вспомогательный контроллер
Auth/enter -> service/authorization
dashboard/get_time -> service/get_time
