+++ Application/Controllers/Controller_service.php
@@ -2,29 +2,39 @@
 /**
  * Created by PhpStorm.
  * User: darevski
- * Date: 30.09.15
- * Time: 14:34
- * @author Darevski
+ * Date: 03.03.16
+ * Time: 19:43
  */
 
 namespace Application\Controllers;
+
 use Application\Core\Controller;
 use Application\Exceptions\UFO_Except;
-use Application\Models\Model_Auth;
-
 /**
- * Контроллер действий связанных с авторизацией пользователей
- * Class Controller_Auth
+ * Class Controller_service
+ * Вспомогательные функции
  * @package Application\Controllers
  */
-class Controller_Auth extends Controller
+class Controller_service extends Controller
 {
+    /**
+     * Выводит json строку со временем на сервере
+     * string now_time - unix timestamp
+     * [state] = 'success'
+     * @api
+     */
+    public function action_get_time(){
+        $date['now_time'] = date("U");
+        $date['state'] = 'success';
+        $this->view->output_json($date);
+    }
+
     /**
      * Функция входа
      * Обеспечивает запись в сессию данных пользователя, для последующей индентификации в системе
      * @throws UFO_Except при не верном пароле вброс исключения с ошибкой не прошедшей авторизации
      */
-    function action_enter(){
+    function action_authorization(){
         if (isset ($_POST['login']) & isset($_POST['password'])) {
             $login = $this->security_variable($_POST['login']);
             $password = $this->security_variable($_POST['password']);
@@ -41,6 +51,4 @@ class Controller_Auth extends Controller
             }
         }
     }
-
-
 }
